feature/YH-55: [Refactor] Updating the ESLint structure and rules
