+++ src/app/App.tsx
@@ -1,12 +1,15 @@
 import { Suspense } from 'react';
-import './styles/normalize.css';
 
-import { Layout } from './components/Layout/Layout';
+import { MainLayout } from './layouts/MainLayout/MainLayout';
+import styles from './styles/App.module.css';
+import './styles/normalize.css';
 
 export const App = () => {
 	return (
 		<Suspense fallback="">
-			<Layout />
+			<div className={styles.container}>
+				<MainLayout />
+			</div>
 		</Suspense>
 	);
 };
