+++ src/app/layouts/MainLayout/MainLayout.module.css
@@ -1,18 +1,12 @@
-@import url("../../styles/App.module.css");
-
 .layout {
     display: grid;
     gap: 0;
     width: 100%;
     height: 100%;
     grid:
-        "logo header" 72px
+        "sidebar header" 72px
         "sidebar content" 1fr
         / 280px 1fr;
-}
-
-.logo {
-    grid-area: logo;
 }
 
 .sidebar {
@@ -21,15 +15,8 @@
 
 .header {
     grid-area: header;
-    background: yellow;
 }
 
 .content {
     grid-area: content;
-}
-
-.container {
-    margin: 0 auto;
-    padding: 0 var(--container-padding);
-    max-width: var(--container-width);
 }
