+++ src/app/styles/App.module.css
@@ -48,3 +48,7 @@ body {
     padding: 0 var(--container-padding);
     max-width: var(--container-width);
 }
+
+*:focus-visible {
+    outline: 4px solid var(--badge-color);
+}
