+++ src/features/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export { SearchInput } from './ui/SearchInput';
