+++ src/features/SearchInput/ui/SearchInput.module.css
