+++ src/pages/Questions/index.ts
