+++ src/shared/ui/AppLogo/AppLogo.tsx
@@ -1,17 +1,11 @@
-import React, { FC, ReactNode } from 'react';
+import { FC } from 'react';
 import { NavLink } from 'react-router-dom';
 
-// eslint-disable-next-line
 import Logo from '@/shared/assets/icons/logo.svg';
 
 import styles from './AppLogo.module.css';
 
-interface AppLogoProps {
-	className: string;
-	children?: ReactNode;
-}
-
-export const AppLogo: FC<AppLogoProps> = () => {
+export const AppLogo: FC = () => {
 	return (
 		<NavLink to="/" className={styles.logo}>
 			<Logo className={styles.image} />
