+++ src/shared/ui/Badge/Badge.tsx
@@ -3,6 +3,7 @@ import React, { FC } from 'react';
 import styles from './Badge.module.css';
 
 interface BadgeProps {
+	/* Проп count используется для динамического отображения числового значения в компоненте Badge */
 	count: number;
 }
