+++ src/shared/ui/Button/Button.tsx
@@ -3,6 +3,7 @@ import React, { FC } from 'react';
 import styles from './Button.module.css';
 
 interface ButtonProps {
+	/*Проп children используется для передачи дочерних элементов */
 	children: string;
 }
