--- src/shared/ui/NavigationSidebar/index.ts
@@ -1,1 +1,0 @@
-export { NavigationSidebar } from './NavigationSidebar';
