--- src/shared/ui/SearchInput/index.ts
@@ -1,1 +1,0 @@
-export { SearchInput } from './SearchInput';
