--- src/widgets/Content/index.ts
@@ -1,1 +1,0 @@
-export { Content } from './ui/Content';
