+++ src/widgets/NavigationSidebar/index.ts
@@ -1,0 +1,2 @@
+export { NavigationSidebar } from './ui/NavigationSidebar/NavigationSidebar';
+export { NavigationSidebarItem } from './ui/NavigationSidebarItem/NavigationSidebarItem';
