+++ src/widgets/NavigationSidebar/ui/NavigationSidebar/NavigationSidebar.module.css
@@ -1,3 +1,4 @@
 .sidebar {
+    height: 100vh;
     background: var(--background-primary);
 }
