+++ src/widgets/NavigationSidebar/ui/NavigationSidebarItem/NavigationSidebarItem.module.css
@@ -3,6 +3,7 @@
     margin-top: 40px;
     padding: 0 20px;
     width: 240px;
+    height: 100%;
     list-style: none;
 }
