+++ src/widgets/SearchSection/index.ts
@@ -1,0 +1,1 @@
+export { SearchSection } from './ui/SearchSection';
