+++ src/widgets/SearchSection/ui/SearchSection.module.css
