--- src/widgets/Sidebar/index.ts
@@ -1,1 +1,0 @@
-export { Sidebar } from './ui/Sidebar';
