--- src/widgets/Sidebar/ui/Sidebar.module.css
@@ -1,3 +1,0 @@
-.sidebar {
-    background: var(--background-primary);
-}
